+++ src/gt4py/next/ffront/decorator.py
@@ -453,33 +453,32 @@ class ProgramWithBoundArgs(Program):
             ) from err
 
         full_args = [*args]
-        full_kwargs = list(kwargs.items())
+        full_kwargs = {**kwargs}
         for index, param in enumerate(self.past_node.params):
             if param.id in self.bound_args.keys():
                 if index < len(full_args):
                     full_args.insert(index, self.bound_args[param.id])
                 else:
-                    pos = list(self.past_node.type.definition.pos_or_kw_args).index(str(param.id))
-                    full_kwargs.insert(pos, (str(param.id), self.bound_args[param.id]))
+                    full_kwargs[str(param.id)] = self.bound_args[param.id]
 
-        return super()._process_args(tuple(full_args), dict(full_kwargs))
+        return super()._process_args(tuple(full_args), full_kwargs)
 
     @functools.cached_property
     def itir(self):
         new_itir = super().itir
         for new_clos in new_itir.closures:
-            for key in self.bound_args.keys():
+            new_args = [ref(inp.id) for inp in new_clos.inputs]
+            for key, value in self.bound_args.items():
                 index = next(
                     index
                     for index, closure_input in enumerate(new_clos.inputs)
                     if closure_input.id == key
                 )
+                new_args[new_args.index(new_clos.inputs[index])] = promote_to_const_iterator(
+                    literal_from_value(value)
+                )
                 new_clos.inputs.pop(index)
-            new_args = [ref(inp.id) for inp in new_clos.inputs]
             params = [sym(inp.id) for inp in new_clos.inputs]
-            for key, value in self.bound_args.items():
-                pos = list(self.past_node.type.definition.pos_or_kw_args).index(key)
-                new_args.insert(pos, promote_to_const_iterator(literal_from_value(value)))
             expr = itir.FunCall(
                 fun=new_clos.stencil,
                 args=new_args,
