+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_execution.py
@@ -898,46 +898,6 @@ def test_docstring(cartesian_case):
     cases.verify(cartesian_case, test_docstring, a, inout=a, ref=a)
 
 
-def test_with_bound_args(cartesian_case):
-    @gtx.field_operator
-    def fieldop_bound_args(a: cases.IField, scalar: int32, condition: bool) -> cases.IField:
-        if not condition:
-            scalar = 0
-        return a + a + scalar
-
-    @gtx.program
-    def program_bound_args(a: cases.IField, scalar: int32, condition: bool, out: cases.IField):
-        fieldop_bound_args(a, scalar, condition, out=out)
-
-    a = cases.allocate(cartesian_case, program_bound_args, "a")()
-    scalar = int32(1)
-    ref = a + a + 1
-    out = cases.allocate(cartesian_case, program_bound_args, "out")()
-
-    prog_bounds = program_bound_args.with_bound_args(scalar=scalar, condition=True)
-    cases.verify(cartesian_case, prog_bounds, a, out, inout=out, ref=ref)
-
-
-def test_with_bound_args_order_args(cartesian_case):
-    @gtx.field_operator
-    def fieldop_args(a: cases.IField, condition: bool, scalar: int32) -> cases.IField:
-        if not condition:
-            scalar = 0
-        return a + scalar
-
-    @gtx.program(backend=cartesian_case.backend)
-    def program_args(a: cases.IField, condition: bool, scalar: int32, out: cases.IField):
-        fieldop_args(a, condition, scalar, out=out)
-
-    a = cases.allocate(cartesian_case, program_args, "a")()
-    scalar = int32(1)
-    ref = a.asnumpy() + scalar
-    out = cases.allocate(cartesian_case, program_args, "out")()
-
-    prog_bounds = program_args.with_bound_args(condition=True)
-    prog_bounds(a=a, scalar=scalar, out=out, offset_provider={})
-    np.allclose(out.asnumpy(), ref)
-
 def test_domain(cartesian_case):
     @gtx.field_operator
     def fieldop_domain(a: cases.IField) -> cases.IField:
